requirements.txt
fastapi==0.74.1
uvicorn==0.17.5
python-dotenv ==0.19.2
aiosqlite==0.17.0
tortoise-orm==0.18.1
